# perf: don't heap allocate a BigInt for every integer that is decoded