@repnop 2024-06-07
it was unrelated, but while investigating failing tests, I noticed a bug in the `DecodeError` display impl which accidentally printed the `self.kind` instead of `self.codec` for the `Codec:` message, so I included that here as well.