@repnop 2024-06-07
> Nice! The performance boost seems to be similar compared what I was doing. I am not sure if it is easy to get faster from integers alone, without changing the logic how constraints are handled. There are a lot of other allocations too, so in general decoding integers alone are not that visible, at least based on my debugging and benchmarking. 

yeah it is kind of a microbenchmark with what I have, I'll try to recreate my original NetSNMP benchmark comparison next week sometime to see how that fairs and perhaps try to stand up something that collects some real-world performance data with SNMP devices I have access to at work to get a better idea of where things could be improved in general as well.

> > but I can perform more performance analysis at some point in the future that specifically looks at the integer decoding alone.
> 
> You can reuse the pure integer bench from here: https://github.com/Nicceboy/rasn/blob/integer-remake/benches/integer.rs

great to know about, thanks! I'll definitely give it a look

> It is significant for Integer type (M2 Pro machine):
> 
> <img width="1190" alt="image" src="https://github.com/librasn/rasn/assets/25366366/a0da706e-1f3b-456f-87d2-60f9d08d9d53">

now those are the numbers I like to see 👀 awesome, thanks for trying it out!